+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -5,6 +5,9 @@ CHANGELOG
 -----
 
  * Removed support to load translation resources from the legacy directories `src/Resources/translations/` and `src/Resources/<BundleName>/translations/`
+ * Removed `ControllerNameParser`.
+ * Removed `ResolveControllerNameSubscriber`
+ * Removed support for `bundle:controller:action` to reference controllers. Use `serviceOrFqcn::method` instead
  * Removed support for PHP templating, use Twig instead
 
 4.4.0